The string_to_integer() function was declared as a real. We already had one of string_to_real(), The function does not appear to be called anywhere in the existing DART code, so no exhaustive testing was performed. I just noticed it when I was trying to merge the rma_trunk updates to the rma_roms branch.
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10617 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: df652954c0980be4cb7fad2a7eb65f7c90a3c40e

+++ utilities/utilities_mod.f90
@@ -2053,7 +2053,7 @@ end function string_to_real
 function string_to_integer(inputstring)
 
 character(len=*), intent(in) :: inputstring
-real(r8)                     :: string_to_integer
+integer                      :: string_to_integer
 
 integer :: io
